Fixed the compiler errors
